[NUI] Split large files
